Removed LED Color Array
Color array was not a great idea. I found a method which loads colors
faster and allows us to take much more advantage of the color creating
capabilities of the TLC5940. Also tried to clean things up again. I
plan to keep doing my best to make this code as efficent as possible.